refactor(network): move get_record code to its own file
